% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom_v2.R
\name{ANCOM}
\alias{ANCOM}
\title{ANCOM main function}
\usage{
ANCOM(
  feature_table,
  meta_data,
  struc_zero = NULL,
  main_var,
  p_adj_method = "BH",
  alpha = 0.05,
  adj_formula = NULL,
  rand_formula = NULL,
  ...
)
}
\arguments{
\item{feature_table}{Data frame representing OTU/SV table with taxa in rows (\code{rownames})
and samples in columns (\code{colnames}). It can be the output value from
\code{feature_table_pre_process.} Note that this is the absolute abundance table, do not
transform it to relative abundance table (where the column totals are equal to 1).}

\item{meta_data}{Data frame of variables. Can be the output value from
\code{feature_table_pre_process.}}

\item{struc_zero}{A matrix consists of 0 and 1s with 1 indicating the taxon is
identified as a structural zero in the corresponding group. Can be the output value
from \code{feature_table_pre_process.}}

\item{main_var}{Character. The name of the main variable of interest. ANCOM v2.1
currently supports categorical \code{main_var.}}

\item{p_adj_method}{Character. Specifying the method to adjust p-values for multiple
comparisons. Default is “BH” (Benjamini-Hochberg procedure).}

\item{alpha}{Level of significance. Default is 0.05.}

\item{adj_formula}{Character string representing the formula for adjustment (see example).}

\item{rand_formula}{Character string representing the formula for random effects in
\code{lme} (see example).}

\item{...}{Additional arguments.}
}
\value{
list
}
\description{
ANCOM main function
}
