% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom_v2.R
\name{feature_table_pre_process}
\alias{feature_table_pre_process}
\title{ANCOM Data Pre-Processing}
\usage{
feature_table_pre_process(
  feature_table,
  meta_data,
  sample_var,
  group_var = NULL,
  out_cut = 0.05,
  zero_cut = 0.9,
  lib_cut,
  neg_lb
)
}
\arguments{
\item{feature_table}{Data frame or matrix representing observed OTU/SV table with taxa
in rows (\code{rownames}) and samples in columns (\code{colnames}). Note that this is the
absolute bundance table, do not transform it to relative abundance table (where the
column totals are equal to 1).}

\item{meta_data}{Data frame or matrix of all variables and covariates of interest.}

\item{sample_var}{Character. The name of column storing sample IDs.}

\item{group_var}{Character. The name of the group indicator. group_var is required for
detecting structural zeros and outliers. For the definitions of different zeros
(structural zero, outlier zero, and sampling zero), please refer to ANCOM-II paper.}

\item{out_cut}{Numerical fraction between 0 and 1. For each taxon, observations with
proportion of mixture distribution less than \code{out_cut} will be detected as outlier
zeros; while observations with proportion of mixture distribution greater than
\code{1 - out_cut} will be detected as outlier values.}

\item{zero_cut}{Numerical fraction between 0 and 1. Taxa with proportion of zeroes
greater than \code{zero_cut} are not included in the analysis.}

\item{lib_cut}{Numeric. Samples with library size less than \code{lib_cut} are not included
in the analysis.}

\item{neg_lb}{Logical. TRUE indicates a taxon would be classified as a structural zero
in the corresponding experimental group using its asymptotic lower bound. More
specifically, \code{neg_lb = TRUE} indicates you are using both criteria stated in section
3.2 of ANCOM-II to detect structural zeros; Otherwise, \code{neg_lb = FALSE} will only use
the equation 1 in section 3.2 of ANCOM-II for declaring structural zeros.}
}
\value{
list
}
\description{
ANCOM Data Pre-Processing
}
\examples{
feature_table_pre_process(
  feature_table = ANCOM:::ecam_table_taxa,
  meta_data = ANCOM:::ecam_sample_metadata,
  sample_var = "Sample.ID",
  group_var = "delivery",
  out_cut = 0.05,
  zero_cut = 0.90,
  lib_cut = 0,
  neg_lb = TRUE
)
}
