% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom_v2.R
\name{outlier_check}
\alias{outlier_check}
\title{Outlier check}
\usage{
outlier_check(x, out_cut)
}
\arguments{
\item{x}{input}

\item{out_cut}{Numerical fraction between 0 and 1. For each taxon, observations with
proportion of mixture distribution less than \code{out_cut} will be detected as outlier
zeros; while observations with proportion of mixture distribution greater than
\code{1 - out_cut} will be detected as outlier values.}
}
\value{
boolean vector
}
\description{
Outlier check
}
